requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.1.0
cycler==0.11.0
dataclasses-json==0.5.14
filelock==3.12.3
Flask==2.3.3
Flask-Cors==4.0.0
fonttools==4.42.1
frozenlist==1.4.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
langchain==0.0.284
langsmith==0.0.33
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numexpr==2.8.5
numpy==1.25.2
openai==0.28.0
opencv-python==4.8.0.76
packaging==23.1
pandas==2.1.0
Pillow==10.0.0
psutil==5.9.5
py-cpuinfo==9.0.0
pydantic==2.3.0
pydantic_core==2.6.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
scipy==1.11.2
seaborn==0.12.2
six==1.16.0
SQLAlchemy==2.0.20
sympy==1.12
tenacity==8.2.3
tiktoken==0.4.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
ultralytics==8.0.173
urllib3==2.0.4
Werkzeug==2.3.7
yarl==1.9.2
zipp==3.16.2
